backend/requirements.txt
pytest==8.3.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
groq==0.9.0
pydantic==2.10.0
python-dotenv==1.0.1
google-generativeai==0.8.0
httpx==0.27.2

